don't make array view functions inline
