+++ src/nanoarrow/array.c
@@ -16,7 +16,6 @@
 // under the License.
 
 #include <errno.h>
-#include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
